requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
Django==4.2.14
djangorestframework==3.15.2
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.2
